Update 12-for-loops.md
The Cumulative Sum challenge creates a variable called 'sum'. 'sum' is a built in function name and using it as a variable is (a) bad practice and (b) breaks the 'Writing Functions' episode which uses the 'sum' function. This edit replaces 'sum' with 'total' to avoid both of those issues.